requirements.txt
click==8.1.3
Flask==2.3.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.3.4
flask-jwt-extended==4.5.2
passlib~=1.7.4
flask-smorest==0.42.0
python-dotenv
pymongo~=4.3.3
jsonpickle~=3.0.1
marshmallow~=3.19.0
discord==2.2.3
